Task: User registration. Added new changes according to the task conditions
