+++ src/main/java/school/faang/users_registration/User.java
@@ -1,4 +1,4 @@
-package school.faang.grouping_users_by_age;
+package school.faang.users_registration;
 
 import lombok.Getter;
 
@@ -9,8 +9,8 @@ import java.util.Set;
 
 @Getter
 public class User {
-    private static final Set<String> PLACE_OF_WORK = Set.of("Yandex", "Sber", "Aston", "VK");
-    private static final Set<String> ADDRESS = Set.of("Moscow", "Saint-Petersburg", "Kazan");
+    private static final Set<String> VALID_JOBS = Set.of("Google", "Uber", "Amazon");
+    private static final Set<String> VALID_ADDRESSES = Set.of("London", "New York", "Amsterdam");
 
     private String userName;
     private int userAge;
@@ -19,19 +19,19 @@ public class User {
 
     public User(String userName, int userAge, String placeOfWork, String address) {
         if (userName == null || userName.isBlank()) {
-            throw new IllegalArgumentException("Имя пользователя не может быть пустым");
+            throw new IllegalArgumentException("User cannot be empty");
         }
 
-        if (userAge <= 0) {
-            throw new IllegalArgumentException("Возраст не может быть нулевым или отрицательным");
+        if (userAge < 18) {
+            throw new IllegalArgumentException("The user must be at least 18 years old");
         }
 
-        if (!PLACE_OF_WORK.contains(placeOfWork)) {
-            throw new IllegalArgumentException("Неверное рабочее место");
+        if (!VALID_JOBS.contains(placeOfWork)) {
+            throw new IllegalArgumentException("Invalid place of work");
         }
 
-        if (!ADDRESS.contains(address)) {
-            throw new IllegalArgumentException("Неверный адрес");
+        if (!VALID_ADDRESSES.contains(address)) {
+            throw new IllegalArgumentException("Invalid address");
         }
         this.userName = userName;
         this.userAge = userAge;
